Add files via upload
These files simulate teh mouvement of two satellites around the hearth. One of themis the hunter and the second one is the target 🎯 
The goal is to take the target through the hunter to avoid the space pollution. If you have any question about it, you can leave me a message and I will respond you asap.